Refactorings on drivers/benchmark dependencies; fixes Dockerfile
- Refactors inner classes in `BenchFlowBenchmark` and `BenchFlowDriver`
(e.g., `RunXml`, `ServiceInfo`…) into separate classes;
- Extracts the logic for starting/stopping BenchFlow services into a
`BenchFlowServices` class;
- Removes outdated resources;
- Updates collectors APIs http methods.
